refactor: common 폴더 추가 및 에러 핸들링 관련 내용 디렉토리 이동
