+++ src/main/java/woowacourse/common/exception/CustomException.java
@@ -1,6 +1,6 @@
-package woowacourse.exception;
+package woowacourse.common.exception;
 
-import woowacourse.exception.dto.ErrorResponse;
+import woowacourse.common.exception.dto.ErrorResponse;
 
 public class CustomException extends RuntimeException {
     private final ErrorResponse errorResponse;
