+++ src/main/java/woowacourse/common/exception/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.exception.dto;
+package woowacourse.common.exception.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
