+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -8,8 +8,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import woowacourse.exception.LoginException;
-import woowacourse.exception.dto.ErrorResponse;
+import woowacourse.common.exception.LoginException;
+import woowacourse.common.exception.dto.ErrorResponse;
 import woowacourse.shoppingcart.domain.customer.Customer;
 import woowacourse.shoppingcart.domain.customer.Password;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
