+++ src/main/java/woowacourse/shoppingcart/ui/ControllerAdvice.java
@@ -2,7 +2,6 @@ package woowacourse.shoppingcart.ui;
 
 import java.util.List;
 import javax.validation.ConstraintViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.validation.BindingResult;
@@ -10,10 +9,7 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import woowacourse.exception.InputFormatException;
-import woowacourse.exception.JoinException;
-import woowacourse.exception.LoginException;
-import woowacourse.exception.dto.ErrorResponse;
+import woowacourse.common.exception.dto.ErrorResponse;
 import woowacourse.shoppingcart.exception.InvalidCartItemException;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 import woowacourse.shoppingcart.exception.InvalidOrderException;
@@ -22,17 +18,6 @@ import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
 
 @RestControllerAdvice
 public class ControllerAdvice {
-
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity handleUnhandledException() {
-        return ResponseEntity.badRequest().body("Unhandled Exception");
-    }
-
-    @ExceptionHandler(EmptyResultDataAccessException.class)
-    public ResponseEntity handle() {
-        return ResponseEntity.badRequest().body("존재하지 않는 데이터 요청입니다.");
-    }
-
     @ExceptionHandler({MethodArgumentNotValidException.class})
     public ResponseEntity handleInvalidRequest(final BindingResult bindingResult) {
         final List<FieldError> fieldErrors = bindingResult.getFieldErrors();
@@ -58,20 +43,5 @@ public class ControllerAdvice {
     })
     public ResponseEntity handleInvalidAccess(final RuntimeException e) {
         return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(LoginException.class)
-    public ResponseEntity<ErrorResponse> handleLoginException(LoginException e) {
-        return ResponseEntity.badRequest().body(e.getErrorResponse());
-    }
-
-    @ExceptionHandler(JoinException.class)
-    public ResponseEntity<ErrorResponse> handleJoinException(JoinException e) {
-        return ResponseEntity.badRequest().body(e.getErrorResponse());
-    }
-
-    @ExceptionHandler(InputFormatException.class)
-    public ResponseEntity<ErrorResponse> handleInputFormatException(InputFormatException e) {
-        return ResponseEntity.badRequest().body(e.getErrorResponse());
     }
 }
