+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.exception.AuthException;
-import woowacourse.exception.JoinException;
+import woowacourse.common.exception.AuthException;
+import woowacourse.common.exception.JoinException;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dto.CustomerResponse;
