+++ src/test/java/woowacourse/shoppingcart/domain/customer/EmailTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.exception.InputFormatException;
+import woowacourse.common.exception.InputFormatException;
 
 class EmailTest {
