+++ src/test/java/woowacourse/shoppingcart/domain/customer/PasswordTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import woowacourse.exception.InputFormatException;
+import woowacourse.common.exception.InputFormatException;
 
 public class PasswordTest {
